Finish building out the model, began work on the server. Added requirements and the xgb joblib file
Finished the model: Changed a few things around to make the code more linear and removed the "run_model" as it isn't necessary and was more so being used for single line testing though that code is all gone anyways.

Server: Taking most of this from a tutorial as I am very new to deploying ML models, will be tuning this to my needs a ton later.

Reqs: Only one I don't know about is scikit-learn and that's becuase I have the sklearn stuff in my files still, will see in the future.

joblib: Seems to have downloaded well, will test once the server aspect is up and running.